// src/shared/api/generated/models/platformCaseIntelligenceSafetyClassDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Retenive Backend Platform
 * CMS, integration, chat and realtime API for Retenive AI Assistant
 * OpenAPI spec version: 0.1.0
 */
import type { PlatformCaseIntelligenceSafetyClassDtoCode } from "./platformCaseIntelligenceSafetyClassDtoCode";
import type { PlatformCaseIntelligenceSafetyClassDtoConsequencesItem } from "./platformCaseIntelligenceSafetyClassDtoConsequencesItem";
import type { PlatformCaseIntelligenceSafetyClassDtoSeverity } from "./platformCaseIntelligenceSafetyClassDtoSeverity";

export interface PlatformCaseIntelligenceSafetyClassDto {
  code: PlatformCaseIntelligenceSafetyClassDtoCode;
  consequences: PlatformCaseIntelligenceSafetyClassDtoConsequencesItem[];
  severity: PlatformCaseIntelligenceSafetyClassDtoSeverity;
}