// src/shared/api/generated/models/platformCaseIntelligenceSafetyPolicyDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Retenive Backend Platform
 * CMS, integration, chat and realtime API for Retenive AI Assistant
 * OpenAPI spec version: 0.1.0
 */
import type { PlatformCaseIntelligenceSafetyPolicyDtoChannelsItem } from "./platformCaseIntelligenceSafetyPolicyDtoChannelsItem";
import type { PlatformCaseIntelligenceSafetyClassDto } from "./platformCaseIntelligenceSafetyClassDto";
import type { PlatformCaseIntelligenceSafetyGateDto } from "./platformCaseIntelligenceSafetyGateDto";

export interface PlatformCaseIntelligenceSafetyPolicyDto {
  /** @maxLength 128 */
  calibratorRevisionId: string;
  /**
   * @minItems 1
   * @maxItems 3
   */
  channels: PlatformCaseIntelligenceSafetyPolicyDtoChannelsItem[];
  /**
   * @minItems 4
   * @maxItems 4
   */
  classes: PlatformCaseIntelligenceSafetyClassDto[];
  /** @maxLength 128 */
  classifierRevisionId: string;
  /** @maxItems 500 */
  gates: PlatformCaseIntelligenceSafetyGateDto[];
  /** @maxLength 128 */
  labelledDatasetRevisionId: string;
  /**
   * @minItems 1
   * @maxItems 50
   */
  locales: string[];
  revisionId: string;
  /** @maxLength 128 */
  sentinelDatasetRevisionId: string;
}