// src/shared/api/generated/models/platformCaseIntelligenceSafetyPolicyDtoChannelsItem.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Retenive Backend Platform
 * CMS, integration, chat and realtime API for Retenive AI Assistant
 * OpenAPI spec version: 0.1.0
 */

export type PlatformCaseIntelligenceSafetyPolicyDtoChannelsItem =
  (typeof PlatformCaseIntelligenceSafetyPolicyDtoChannelsItem)[keyof typeof PlatformCaseIntelligenceSafetyPolicyDtoChannelsItem];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PlatformCaseIntelligenceSafetyPolicyDtoChannelsItem = {
  TEXT: "TEXT",
  VOICE: "VOICE",
  TELEGRAM: "TELEGRAM",
} as const;